Yet more annotations.
This commit adds (skip) annotations to the @parent field of all
the C constructors, so that they don't do the ownership trick
and can be annotated (transfer full) in every case. This is actually
a bit pointless since all the constructers use varargs so are
disabled in the introspection GIR anyway...

But hurrah! No more warnings from g-ir-scanner! (Apart from one that
can't be fixed)